+++ src/components/componentBarLineChart/CHANGELOG.md
@@ -1,5 +1,11 @@
 # component-bar-line-chart
 
+## 0.0.3
+
+### Patch Changes
+
+- fixing error on displaying weekly statistics calories burned
+
 ## 0.0.2
 
 ### Patch Changes